% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dists.R
\name{makeStatList}
\alias{makeStatList}
\title{Internal SuppDists function.}
\usage{
makeStatList(head, mn, med, var, mod, third, fourth, dig)
}
\description{
This is not to be called by the user.
}
\keyword{distribution}
\keyword{internal}
