% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuppDists-package.R
\docType{package}
\name{SuppDists-package}
\alias{SuppDists-package}
\alias{SuppDists}
\title{Supplementary distributions and two random number generators.}
\description{
The package includes ten distributions supplementing those built into \R:
}
\details{
\itemize{
\item \code{\link[=dinvGauss]{dinvGauss()}} The inverse Gaussian and Wald distributions
\item \code{\link[=dKruskalWallis]{dKruskalWallis()}} Kruskall-Wallis distribution
\item \code{\link[=dKendall]{dKendall()}} Kendall's Tau
\item \code{\link[=dFriedman]{dFriedman()}} Friedman's chi squared
\item \code{\link[=dSpearman]{dSpearman()}} Spearman's rho
\item \code{\link[=dmaxFratio]{dmaxFratio()}} The maximum F-ratio distribution
\item \code{\link[=dPearson]{dPearson()}} Pearson product moment correlation coefficient
\item \code{\link[=dJohnson]{dJohnson()}} Johnson distributions
\item \code{\link[=dNormScore]{dNormScore()}} normal scores
\item \code{\link[=dghyper]{dghyper()}} Generalized hypergeometric distributions (Kemp and Kemp)
}

In addition, two random number generators of George Marsaglia:
\itemize{
\item \code{\link[=rMWC1019]{rMWC1019()}} A very long period pseudo-random generator
\item \code{\link[=rziggurat]{rziggurat()}} The Ziggurat normal and exponential generator
}
}
\author{
Original author: Bob Wheeler \email{bwheelerg@gmail.com};
Maintainer: Andrie de Vries \email{apdevries@gmail.com}
}
\keyword{package}
